+++ back/recipe/urls.py
@@ -1,6 +1,8 @@
+from .views import (RecipeList,
+                    RecipeDetail)
 from django.urls import path
-from .views import RecipeListView
 
 urlpatterns = [
-    path('', RecipeListView.as_view(), name='home'),
+    path('recipes/', RecipeList.as_view()),
+    path('recipe/<int:id>/', RecipeDetail.as_view()),
 ]
